refactor: move URLValidationError to plugin_errors module
Relocates URLValidationError (and its validator-message whitelist) from
utils.security_utils into utils.plugin_errors so cross-process callers —
most notably the subprocess worker that reconstructs exceptions by name —
can reference the type without importing security_utils.

SonarCloud's S7788 architecture rule flags imports of security_utils from
refresh_task.worker and the two test files that pin the cross-process
contract. Moving the exception to the plugin_errors module satisfies the
rule while keeping security_utils as the stable public surface for code
that actually calls validate_url / validate_file_path. The class is still
re-exported from security_utils for backward compatibility.

No behaviour change — the exception hierarchy, the safe_message() whitelist,
and the validator error messages are all preserved verbatim.
